Fix #25: Use `__unsafe_unretained` for the NativeLedgerBridge
Obj-C++ seems to have some issues with `__weak` in a C++ class. During the `bridge_`'s dealloc, it's pointer will be dangling instead of nil.

Considering that the NativeLedgerClient object will only ever be alive while the `BATBraveLedger` object is alive we can use `__unsafe_unretained` and delete the ledger/ledger client as usual in the `BATBraveLedger.dealloc`