Add `-Weverything` warning configuration on `BraveRewards` target
One note: `-Wno-nullable-to-nonnull-conversion` is added to ignored rules, but it should be properly handled at some point.

Obj-C does not have any nice `if-let`/`guard-let` to "unwrap" its nullable properties, so you have to manually cast it to get rid of the warning which is not ideal. Also, `std::string` has a default (empty string), therefore we can assume anything that comes from C++ side is nonnull.